+++ lib/lamby.rb
@@ -9,6 +9,7 @@ require 'lamby/rack_rest'
 require 'lamby/rack_http'
 require 'lamby/debug'
 require 'lamby/runner'
+require 'lamby/command'
 require 'lamby/handler'
 
 if defined?(Rails)